Epic 5.0

+++ hr/src/main/java/com/company/hr/core/dto/EmployeeCompensationResponse.java
@@ -3,13 +3,14 @@ package com.company.hr.core.dto;
 import lombok.Getter;
 import lombok.Setter;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 @Setter
 @Getter
 public class EmployeeCompensationResponse {
 
-    private Double basicSalary;
+    private BigDecimal basicSalary;
     private Double overTimeRate;
     private String schedule;
     private List<AllowanceResponse> allowanceResponses;
